requirements.txt
selenium
webdriver-manager
requests
pandas
openpyxl
python-dotenv
google-api-python-client
google-auth-oauthlib
google-auth-httplib2
msal
xlrd
